+++ HISTORY.md
@@ -25,6 +25,8 @@
     * new "get graphic effect" reporter
     * new "get pen attribute" reporter
     * new "write" command in pen category (used to be "label" in tools)
+    * new "map","keep", "combine" and "for each" primitives in list category
+    * new "for" loop and "if then else" reporter primitives in the Control category
     * added "neg", "lg" (log2) and "2^" selectors to monadic function reporter in Operators
     * added "^" reporter (power of) in the Operators category
     * added "width" and "height" as attribute selectors of the OF primitive for the stage
@@ -67,8 +69,26 @@
     * German
     * French
 
+### 2019-04-24
+* updated German translation (for new HOF prims)
+
+### 2019-04-24
+* Threads, Objects: new "combine" primitive in list category
+* Threads: added type-assertions for the new HOF prims
+* Threads, Objects: new "for" loop primitive in Control category
+* Threads, Objects: new "if then else" reporter primitive in Control category
+
+### 2019-04-23
+* Threads: fixed JS stack overflow issue for MAP primitive
+* Threads: new "map" and "for each" primitives in list category
+* Threads: new "keep" primitive in list category
+
+### 2019-04-22
+* Threads: fixed variable binding for "arguments", turned dictionary key into a Symbol
+
 ### 2019-04-15
 * Catalan translation update
+
 ### 2019-04-12
 * Objects: enabled text-variables as inputs for graphic effects / pen attributes
 * updated amination library with graphic effects and audio attributes